--- medical-card/app/page.tsx
@@ -1,14 +1,0 @@
-
-import Link from 'next/link';
-
-export default function Home() {
-  return (
-    <>
-    
-<button className="p-6 bg-sky-500 hover:bg-sky-700 text-white font-bold py-2 px-4 rounded">
- <Link  href="about">Read about our Project!</Link>
-</button>
-</>     
-  )
-
-}
